requirements.txt
fastapi==0.111.0
pydantic==1.9.0
sqlalchemy==2.0.31
psycopg2==24.1.2
passlib[bcrypt]
PyJWT


